Add the HTML, CSS code, an image and README.md
Add all the HTML code, all the CSS code, the image math-chalkboard.jpg
and the README.md description.

The HTML and CSS code are sufficient to pass the test. Both basically
follow the FCC example, but add a bit of personal touch. The image is
used in the background to improve the page design. The content on
README.md will help people understand the objectives of the project.